--- src/app/components/OptimizelyPageMetrics/PageViewTracking/experiments.tsx
@@ -1,4 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-
-// TODO - any running serverside and client side experiments which collect metric PageCompleteTracking
-export const experiments = [''];
