+++ src/app/components/OptimizelyPageMetrics/ScrollDepthTracking/index.tsx
@@ -1,7 +1,8 @@
 import useOptimizelyScrollDepth from '#hooks/useOptimizelyScrollDepth';
+import { experimentsForScrollDepthTracking as experiments } from '../experimentsForPageMetrics';
 
 const ScrollDepthTracking = () => {
-  useOptimizelyScrollDepth();
+  useOptimizelyScrollDepth(experiments);
   return null;
 };
