+++ src/app/components/OptimizelyPageMetrics/experimentsForPageMetrics.tsx
@@ -1,0 +1,17 @@
+// TODO - any running serverside and client side experiments which collect metric PageCompleteTracking
+export const experimentsForPageCompleteTracking: string[] = [
+  'dummy_experiment',
+  'dummy_experiment_1',
+];
+
+// TODO - any running serverside and client side experiments which collect metric PageViewTracking
+export const experimentsForPageViewTracking: string[] = [
+  'dummy_experiment',
+  'dummy_experiment_1',
+];
+
+// TODO - any running serverside and client side experiments which collect metric ScrollDepthTracking
+export const experimentsForScrollDepthTracking: string[] = [
+  'dummy_experiment',
+  'dummy_experiment_1',
+];
